chore : rename package from 'piece' to 'pieces'
