+++ app/src/test/java/org/chess/domain/board/BoardTest.java
@@ -1,6 +1,6 @@
 package org.chess.domain.board;
 
-import org.chess.domain.piece.Pawn;
+import org.chess.domain.pieces.Pawn;
 import org.junit.jupiter.api.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
